Update SessionController.php
This is a kind of counter-intuitive change, but the field is named "email" but the system actually treats it as a username when it looks it up in the database. Kind of a dirty quick fix for now. Perhaps somebody can go around and renamed all instances of `$_POST['email']` to $_POST['username'] etc.